Move class `Debouncer` to its own file
